MAINT: move _decorator package to util
