--- dtoolkit/_decorator/__init__.py
@@ -1,1 +1,0 @@
-from dtoolkit._decorator.warning import warning  # noqa
