+++ dtoolkit/util/__init__.py
@@ -1,2 +1,2 @@
-from dtoolkit.util.multi_if_else import multi_if_else  # noqa
-from dtoolkit.util.snake_to_camel import snake_to_camel  # noqa
+from dtoolkit.util.generic import multi_if_else  # noqa
+from dtoolkit.util.generic import snake_to_camel  # noqa
