+++ dtoolkit/util/generic/__init__.py
@@ -1,0 +1,2 @@
+from .multi_if_else import multi_if_else  # noqa
+from .snake_to_camel import snake_to_camel  # noqa
