+++ dtoolkit/util/generic/multi_if_else.py
