+++ dtoolkit/util/generic/snake_to_camel.py
